+++ src-opam/dockerfile_distro.ml
@@ -37,6 +37,7 @@ type status = [
 type arch = [
   | `X86_64
   | `Aarch64
+  | `Ppc64le
 ] [@@deriving sexp]
 
 let distros = [