+++ CHANGES.md
@@ -1,6 +1,7 @@
 v5.0.0
 ------
 
+- Add Ppc64le architecture.
 - Install the Bubblewrap sandboxing tool in all distributions.
 - Ensure that X11 is available in the containers so that the
   OCaml Graphics module is available (#8 via @kit-ty-kate)